asset optimizer refactoring
